+++ RELEASE-NOTES.md
@@ -4,6 +4,8 @@
 
 **Enhancements**
 - **Grid** - `equal height` and `equal width` now work without `row` wrappers
+- **Visibility/Sticky** - Visibility now uses pub/sub pattern to greatly improve scroll performance when attaching multiple events
+- **Visibility/Sticky** - Visibility and sticky now refresh automatically after page content loading to deal with changes in position from images loading
 
 **Bugs**
 - **All Modules** - Performance logging now delays 500ms instead of 100ms for console logging to ensure all logs are captured in one group